pyproject.toml
[project]
name = "nw-london-health-pipeline"
version = "0.1.0"
description = "Auto-updating data pipeline for the NW London health map. Outputs Parquet for Power BI + JSON for Leaflet."
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31",
    "pandas>=2.2",
    "pyarrow>=15.0",         # Parquet I/O
    "pyyaml>=6.0",
    "shapely>=2.0",          # geometry + STRtree spatial index
    "pyproj>=3.6",           # BNG <-> WGS84
    "pyshp>=2.3",            # ESRI shapefile reader
    "tqdm>=4.66",
    "rich>=13.7",            # nicer logging
    "click>=8.1",            # CLI
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]
raster = [
    "rasterio>=1.3",
    "rioxarray>=0.15",
]

[project.scripts]
pipeline = "pipeline.src.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["pipeline*"]
exclude = ["pipeline.tests*", "pipeline.conf*"]

[tool.setuptools.package-data]
"pipeline.src.fetchers.outcomes" = ["*.yml"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["pipeline/tests"]
